+++ app/controllers/solid_queue_guard/mission_control/base_controller.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module SolidQueueGuard
+  module MissionControl
+    class BaseController < ::MissionControl::Jobs::ApplicationController
+      helper DashboardHelper
+    end
+  end
+end