+++ test/dummy/config/routes.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
 Rails.application.routes.draw do
-  mount SolidQueueGuard::Engine, at: "/solid_queue_guard"
+  mount MissionControl::Jobs::Engine, at: '/jobs'
+  mount SolidQueueGuard::Engine, at: '/solid_queue_guard'
 end